+++ archunit/src/main/java/com/tngtech/archunit/core/domain/JavaClass.java
@@ -658,11 +658,6 @@ public final class JavaClass
     }
 
     @Override
-    public Set<JavaClass> getAllInvolvedRawTypes() {
-        return ImmutableSet.of(getBaseComponentType());
-    }
-
-    @Override
     public void traverseSignature(SignatureVisitor visitor) {
         SignatureTraversal.from(visitor).visitClass(this);
     }
