+++ archunit/src/main/java/com/tngtech/archunit/core/importer/DomainBuilders.java
@@ -88,7 +88,6 @@ import static com.tngtech.archunit.core.domain.JavaConstructor.CONSTRUCTOR_NAME;
 import static com.tngtech.archunit.core.domain.properties.HasName.Utils.namesOf;
 import static java.util.Collections.emptyList;
 import static java.util.stream.Collectors.joining;
-import static java.util.stream.Collectors.toSet;
 
 @Internal
 @SuppressWarnings("UnusedReturnValue")
@@ -1215,14 +1214,6 @@ public final class DomainBuilders {
         @Override
         public JavaClass toErasure() {
             return type.toErasure();
-        }
-
-        @Override
-        public Set<JavaClass> getAllInvolvedRawTypes() {
-            return Stream.concat(
-                    type.getAllInvolvedRawTypes().stream(),
-                    typeArguments.stream().map(JavaType::getAllInvolvedRawTypes).flatMap(Set::stream)
-            ).collect(toSet());
         }
 
         @Override
